refactor: 로컬스토리지 관련함수 유틸로 분리, 모델이 상태값 갖도록 변경
