+++ src/js/controllers/elementValidator.js
@@ -1,3 +1,7 @@
+import { KEY_VIDEOS_WATCHING, KEY_VIDEOS_WATCHED } from '../constants.js';
+
+import { getListByKey } from '../utils/localStorage.js';
+
 export const isModalOpen = (currentTarget) => {
   return currentTarget.querySelector('.modal').classList.contains('open');
 };
@@ -8,4 +12,14 @@ export const isModalDimmedArea = (target) => {
 
 export const isModalCloseButton = (target) => {
   return target.closest('.js-modal-close-button');
+};
+
+export const isSavedVideo = (targetId) => {
+  const videosWatching = getListByKey(KEY_VIDEOS_WATCHING);
+  const videosWatched = getListByKey(KEY_VIDEOS_WATCHED);
+
+  return (
+    videosWatching.some((video) => video.videoId === targetId) ||
+    videosWatched.some((video) => video.videoId === targetId)
+  );
 };
