+++ src/js/models/SearchModel.js
@@ -1,49 +1,73 @@
-import { setListToLocalStorage, getListFromLocalStorage } from '../utils/localStorage.js';
+import {
+  MAX_RECENT_KEYWORD_COUNT,
+  KEY_VIDEOS_WATCHING,
+  KEY_VIDEOS_WATCHED,
+  KEY_RECENT_KEYWORDS,
+} from '../constants.js';
+
+import { getListByKey, insertItemByKey, insertItemAtFirstByKey, deleteLastItemByKey } from '../utils/localStorage.js';
 
 /*
 로컬스토리지 데이터 저장구조
   {
-    videosToWatch: [ { videoId: 'TVUNi0zi1yw', ... }, ... ],
+    videosWatching: [ { videoId: 'TVUNi0zi1yw', ... }, ... ],
     videosWatched: [ { videoId: '0nxsS4B85E8', ... }, ... ],
     recentKeywords: [ 'keyword1', 'keyword2', 'keyword3' ],
   }
 */
 
 export default class SearchModel {
-  getListByKey(key) {
+  constructor(keyword = '') {
+    this.init(keyword);
+  }
+
+  init(keyword) {
+    this.nextPageToken = '';
+    this.totalSearchResult = [];
+    this.keyword = keyword;
+    this.saveKeyword();
+  }
+
+  setNextPageToken(token) {
+    this.nextPageToken = token;
+  }
+
+  saveKeyword() {
+    if (this.keyword === '') {
+      return;
+    }
+
+    const recentKeywords = getListByKey(KEY_RECENT_KEYWORDS);
+
+    if (recentKeywords.includes(this.keyword)) {
+      return;
+    }
+    if (recentKeywords.length === MAX_RECENT_KEYWORD_COUNT) {
+      deleteLastItemByKey(KEY_RECENT_KEYWORDS);
+    }
+    insertItemAtFirstByKey(KEY_RECENT_KEYWORDS, this.keyword);
+  }
+
+  getRecentKeywords() {
     try {
-      return JSON.parse(getListFromLocalStorage(key)) || [];
+      return getListByKey(KEY_RECENT_KEYWORDS);
     } catch (e) {
-      return [];
+      return '';
     }
   }
 
-  setListByKey(key, list) {
-    setListToLocalStorage(key, JSON.stringify(list));
+  addNewSearchResult(newSearchResult) {
+    this.totalSearchResult.push(...newSearchResult);
   }
 
-  insertItemByKey(key, item) {
-    const list = this.getListByKey(key);
-    list.push(item);
-    this.setListByKey(key, list);
+  saveVideo(targetId) {
+    const targetVideo = this.totalSearchResult.find((video) => video.videoId === targetId);
+
+    targetVideo.isSaved = true;
+    insertItemByKey(KEY_VIDEOS_WATCHING, targetVideo);
   }
 
-  insertItemAtFirstByKey(key, item) {
-    const list = this.getListByKey(key);
-    list.unshift(item);
-    this.setListByKey(key, list);
-  }
-
-  deleteLastItemByKey(key) {
-    const list = this.getListByKey(key);
-
-    this.setListByKey(key, list.slice(0, list.length - 1));
-  }
-
-  deleteTargetItemByKey(key, secondKey, target) {
-    const list = this.getListByKey(key);
-    const filteredList = list.filter((item) => item[secondKey] !== target[secondKey]);
-
-    this.setListByKey(key, filteredList);
+  getSavedVideoCount() {
+    return getListByKey(KEY_VIDEOS_WATCHING).length + getListByKey(KEY_VIDEOS_WATCHED);
   }
 }
